removes unnecessary specific file
